Improve bool'ness (and add explicit static_cast<int>() for printf like operations to be compatible of a bool being a C++ object) (https://trac.osgeo.org/gdal/ticket/6325)
git-svn-id: https://svn.osgeo.org/gdal/trunk@33093 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogr_srs_esri.cpp
@@ -1863,7 +1863,7 @@ OGRErr OGRSpatialReference::morphFromESRI()
                             exportToWkt ( &pszOtherValue );
                             CPLDebug( "OGR_ESRI", 
                                       "morphFromESRI() got SRS %s, matching: %d", 
-                                      pszOtherValue, bIsSame );
+                                      pszOtherValue, static_cast<int>(bIsSame) );
                             CPLFree( pszOtherValue );
                             delete poSRSTemp2;
